folder structure and module refactor
