+++ routes/apiRoutes/index.js
@@ -1,0 +1,5 @@
+const router = require("express").Router();
+
+router.use(require("./animalRoutes"))
+
+module.exports = router;
